requirements.txt
numpy==1.19.4
torch==1.8.0
pandas
scikit_learn
pyYAML